Moving board into same directory
